Refactored RapidsErrorUtils
