+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/execution/python/BatchGroupUtils.scala
@@ -31,6 +31,7 @@ import org.apache.spark.sql.execution.SparkPlan
 import org.apache.spark.sql.rapids.execution.python.shims.{GpuPythonArrowOutput, GpuPythonRunnerBase}
 import org.apache.spark.sql.rapids.shims.DataTypeUtilsShim
 import org.apache.spark.sql.vectorized.ColumnarBatch
+
 /**
  * A helper class to pack the group related items for the Python input.
  *
