+++ sql-plugin/src/main/spark330db/scala/org/apache/spark/sql/rapids/shims/RapidsErrorUtils.scala
@@ -20,71 +20,4 @@
 spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql.rapids.shims
 
-import org.apache.spark.SparkDateTimeException
-import org.apache.spark.sql.catalyst.trees.{Origin, SQLQueryContext}
-import org.apache.spark.sql.errors.QueryExecutionErrors
-import org.apache.spark.sql.internal.SQLConf
-import org.apache.spark.sql.types.{DataType, Decimal, DecimalType}
-
-object RapidsErrorUtils extends RapidsErrorUtilsFor330plus {
-
-  def mapKeyNotExistError(
-      key: String,
-      keyType: DataType,
-      origin: Origin): NoSuchElementException = {
-    throw new UnsupportedOperationException(
-      "`mapKeyNotExistError` has been removed since Spark 3.4.0. "
-    )
-  }
-
-  def invalidArrayIndexError(
-      index: Int,
-      numElements: Int,
-      isElementAtF: Boolean = false,
-      context: SQLQueryContext = null): ArrayIndexOutOfBoundsException = {
-    if (isElementAtF) {
-      QueryExecutionErrors.invalidElementAtIndexError(index, numElements, context)
-    } else {
-      QueryExecutionErrors.invalidArrayIndexError(index, numElements, context)
-    }
-  }
-
-  def arithmeticOverflowError(
-      message: String,
-      hint: String = "",
-      errorContext: SQLQueryContext = null): ArithmeticException = {
-    QueryExecutionErrors.arithmeticOverflowError(message, hint, errorContext)
-  }
-
-  def cannotChangeDecimalPrecisionError(
-      value: Decimal,
-      toType: DecimalType,
-      context: SQLQueryContext = null): ArithmeticException = {
-    QueryExecutionErrors.cannotChangeDecimalPrecisionError(
-      value, toType.precision, toType.scale, context
-    )
-  }
-
-  def overflowInIntegralDivideError(context: SQLQueryContext = null): ArithmeticException = {
-    QueryExecutionErrors.arithmeticOverflowError(
-      "Overflow in integral divide", "try_divide", context
-    )
-  }
-
-  def sparkDateTimeException(infOrNan: String): SparkDateTimeException = {
-    // These are the arguments required by SparkDateTimeException class to create error message.
-    val errorClass = "CAST_INVALID_INPUT"
-    val messageParameters = Map("expression" -> infOrNan, "sourceType" -> "DOUBLE",
-      "targetType" -> "TIMESTAMP", "ansiConfig" -> SQLConf.ANSI_ENABLED.key)
-    SparkDateTimeExceptionShims.newSparkDateTimeException(errorClass, messageParameters,
-      Array.empty, "")
-  }
-
-  def sqlArrayIndexNotStartAtOneError(): RuntimeException = {
-    QueryExecutionErrors.elementAtByIndexZeroError(context = null)
-  }
-  
-  override def intervalDivByZeroError(origin: Origin): ArithmeticException = {
-    QueryExecutionErrors.intervalDividedByZeroError(origin.context)
-  }
-}
+object RapidsErrorUtils extends RapidsErrorUtilsBase
